Split SignificantTermsSignificanceScoreTests in two. Groovy-related tests (File/inline and indexed scripts) are retained in lang-groovy plugin’s test framework. All other non-groovy test methods relating to core significance scores and native plugins moved to new class in core-tests.
